Jdbc update one by unique filter attr. OBS: updated Float type Real requires updated sbt-molecule (`git pull` and `sbt publishLocal`) and re-generation of boilerplate code (run `sbt clean compile -Dmolecule=true`) Renaming "Unique" namespace to "Uniques" to avoid clash with h2 reserved keyword.

+++ sql/core/shared/src/main/scala/molecule/sql/core/query/LambdasOne.scala
@@ -45,7 +45,7 @@ trait LambdasOne extends ResolveBase { self: Base =>
   private lazy val sql2oneStringOrNull    : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getString(n); if (row.wasNull()) null else v }
   private lazy val sql2oneIntOrNull       : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getInt(n); if (row.wasNull()) null else v }
   private lazy val sql2oneLongOrNull      : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getLong(n); if (row.wasNull()) null else v }
-  private lazy val sql2oneFloatOrNull     : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getDouble(n); if (row.wasNull()) null else v }
+  private lazy val sql2oneFloatOrNull     : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getFloat(n); if (row.wasNull()) null else v }
   private lazy val sql2oneDoubleOrNull    : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getDouble(n); if (row.wasNull()) null else v }
   private lazy val sql2oneBooleanOrNull   : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getBoolean(n); if (row.wasNull()) null else v }
   private lazy val sql2oneBigIntOrNull    : (Row, Int) => Any = { (row: Row, n: Int) => val v = row.getBigDecimal(n); if (row.wasNull()) null else v.toBigInteger }
